+++ R/AcqOptimzerRandomSearch.R
@@ -1,5 +1,18 @@
 #' @title Random Search Acquisition Function Optimizer
 #'
+#' @include AcqOptimizer.R mlr_acqoptimizers.R
+#'
+#' @description
+#' Random search acquisition function optimizer.
+#' By default, it samples `100 * D^2` random points in the search space, where `D` is the dimension of the search space.
+#' The point with the highest acquisition value is returned.
+#'
+#' @section Parameters:
+#' \describe{
+#' \item{`n_evals`}{`integer(1)`\cr
+#'   Number of random points to sample.
+#'   Default is `100 * D^2`, where `D` is the dimension of the search space.}
+#' }
 #' @export
 AcqOptimizerRandomSearch = R6Class("AcqOptimizerRandomSearch",
   inherit = AcqOptimizer,
@@ -12,40 +25,9 @@ AcqOptimizerRandomSearch = R6Class("AcqOptimizerRandomSearch",
     initialize = function(acq_function = NULL) {
       self$acq_function = assert_r6(acq_function, "AcqFunction", null.ok = TRUE)
       param_set = ps(
-        max_fevals = p_int(lower = 1, init = 1000L)
-        # maxeval          = p_int(lower = 1, init = 1000L, special_vals = list(-1)),
-        # fnscale          = p_dbl(default = 1),
-        # #maxit            = p_int(lower  = 1L),
-        # stopfitness      = p_dbl(default = -Inf),
-        # keep.best        = p_lgl(default = TRUE),
-        # sigma            = p_uty(default = 0.5),
-        # mu               = p_int(lower = 1L),
-        # lambda           = p_int(lower = 1L),
-        # weights          = p_uty(),
-        # damps            = p_dbl(),
-        # cs               = p_dbl(),
-        # ccum             = p_dbl(),
-        # ccov.1           = p_dbl(lower = 0),
-        # ccov.mu          = p_dbl(lower = 0),
-        # diag.sigma       = p_lgl(default = FALSE),
-        # diag.eigen       = p_lgl(default = FALSE),
-        # diag.pop         = p_lgl(default = FALSE),
-        # diag.value       = p_lgl(default = FALSE),
-        # stop.tolx        = p_dbl(), # undocumented stop criterion
-        # restart_strategy = p_fct(levels = c("none", "ipop"), init = "none"),
-        # n_restarts       = p_int(lower = 0L, init = 0L),
-        # population_multiplier = p_int(lower = 1, init = 2L)
-        # start_values  = p_fct(default = "random", levels = c("random", "center", "custom")),
-        # start         = p_uty(default = NULL, depends = start_values == "custom"),
-        # n_candidates = p_int(lower = 1, default = 1L),
-        # logging_level = p_fct(levels = c("fatal", "error", "warn", "info", "debug", "trace"), default = "warn"),
-        # warmstart = p_lgl(default = FALSE),
-        # warmstart_size = p_int(lower = 1L, special_vals = list("all")),
-        # skip_already_evaluated = p_lgl(default = TRUE),
-        # catch_errors = p_lgl(default = TRUE)
+        n_evals = p_int(lower = 1),
+        catch_errors = p_lgl(init = TRUE)
       )
-      # ps$values = list(n_candidates = 1, logging_level = "warn", warmstart = FALSE, skip_already_evaluated = TRUE, catch_errors = TRUE)
-      # ps$add_dep("warmstart_size", on = "warmstart", cond = CondEqual$new(TRUE))
       private$.param_set = param_set
     },
 
@@ -60,9 +42,27 @@ AcqOptimizerRandomSearch = R6Class("AcqOptimizerRandomSearch",
       constants = self$acq_function$constants$values
       direction = self$acq_function$codomain$direction
 
-      xdt = generate_design_random(self$acq_function$domain, n = pv$max_fevals)$data
+      if (is.null(pv$n_evals)) {
+        pv$n_evals = 100 * self$acq_function$domain$length^2
+      }
 
-      ys = mlr3misc::invoke(fun, xdt = xdt, .args = constants)[[1]]
+
+      xdt = generate_design_random(self$acq_function$domain, n = pv$n_evals)$data
+
+      optimize = function() {
+        mlr3misc::invoke(fun, xdt = xdt, .args = constants)[[1]]
+      }
+
+      if (pv$catch_errors) {
+        tryCatch({
+          ys = optimize()
+        }, error = function(error_condition) {
+          lg$warn(error_condition$message)
+          stop(set_class(list(message = error_condition$message, call = NULL), classes = c("acq_optimizer_error", "mbo_error", "error", "condition")))
+        })
+      } else {
+        ys = optimize()
+      }
 
       id = if (direction == 1) which.min(ys) else which.max(ys)
       x = xdt[id, ]
@@ -70,34 +70,14 @@ AcqOptimizerRandomSearch = R6Class("AcqOptimizerRandomSearch",
 
       set(x, j = self$acq_function$codomain$ids(), value = y)
       x
-    },
-
-    #' @description
-    #' Reset the acquisition function optimizer.
-    #'
-    #' Currently not used.
-    reset = function() {
-
     }
   ),
 
   active = list(
     #' @template field_print_id
     print_id = function(rhs) {
-      if (missing(rhs)) {
-        paste0("(", class(self$optimizer)[1L], " | ", class(self$terminator)[1L], ")")
-      } else {
-        stop("$print_id is read-only.")
-      }
-    },
-
-    #' @field param_set ([paradox::ParamSet])\cr
-    #'   Set of hyperparameters.
-    param_set = function(rhs) {
-      if (!missing(rhs) && !identical(rhs, private$.param_set)) {
-        stop("$param_set is read-only.")
-      }
-      private$.param_set
+      assert_ro_binding(rhs)
+      "(OptimizerRandomSearch)"
     }
   ),
 
@@ -116,3 +96,4 @@ AcqOptimizerRandomSearch = R6Class("AcqOptimizerRandomSearch",
   )
 )
 
+mlr_acqoptimizers$add("random_search", AcqOptimizerRandomSearch)